# opa: do not inline deny (IAM v2 performance)