… components/authz-service/engine/opa/policy.bindata.go
… vendor/github.com/open-policy-agent/opa/ast/builtins.go
… vendor/github.com/open-policy-agent/opa/ast/check.go
… vendor/github.com/open-policy-agent/opa/ast/compare.go
… vendor/github.com/open-policy-agent/opa/ast/compile.go
… vendor/github.com/open-policy-agent/opa/ast/doc.go
… vendor/github.com/open-policy-agent/opa/ast/parser.go
… vendor/github.com/open-policy-agent/opa/ast/parser_internal.go
… vendor/github.com/open-policy-agent/opa/ast/policy.go
… vendor/github.com/open-policy-agent/opa/ast/term.go
… vendor/github.com/open-policy-agent/opa/ast/transform.go
… vendor/github.com/open-policy-agent/opa/ast/varset.go
… vendor/github.com/open-policy-agent/opa/ast/visit.go
… vendor/github.com/open-policy-agent/opa/internal/compiler/wasm/opa/opa.go
… vendor/github.com/open-policy-agent/opa/internal/compiler/wasm/wasm.go
… vendor/github.com/open-policy-agent/opa/internal/ir/ir.go
… vendor/github.com/open-policy-agent/opa/internal/ir/walk.go
… vendor/github.com/open-policy-agent/opa/internal/planner/functrie.go
… vendor/github.com/open-policy-agent/opa/internal/planner/planner.go
… vendor/github.com/open-policy-agent/opa/internal/wasm/instruction/control.go
… vendor/github.com/open-policy-agent/opa/internal/wasm/module/module.go
… vendor/github.com/open-policy-agent/opa/metrics/metrics.go
… vendor/github.com/open-policy-agent/opa/rego/rego.go
… vendor/github.com/open-policy-agent/opa/topdown/builtins.go
… vendor/github.com/open-policy-agent/opa/topdown/cache.go
… vendor/github.com/open-policy-agent/opa/topdown/copypropagation/copypropagation.go
… vendor/github.com/open-policy-agent/opa/topdown/errors.go
… vendor/github.com/open-policy-agent/opa/topdown/eval.go
… vendor/github.com/open-policy-agent/opa/topdown/input.go
… vendor/github.com/open-policy-agent/opa/topdown/instrumentation.go