### :nut_and_bolt: Description

This modifies how OPA evaluates policies in partial evaluation -- i.e. for every update of IAM v2 related data (policies + roles). We've found before that it's **annoyingly slow**. It should be less so now, if we can trust our benchmarks:

```
name                                                                                              old time/op    new time/op    delta
InitPartialResultV2/default_policies-8                                                               1.03s ± 3%     0.12s ± 3%  -88.51%  (p=0.000 n=10+10)
InitPartialResultV2/IsAuthorized(true)_with_the_result_of_BenchmarkInitPartialResultV2-8             279µs ± 7%     282µs ± 1%   +1.01%  (p=0.035 n=9+10)
InitPartialResultV2/IsAuthorized(false)_with_the_result_of_BenchmarkInitPartialResultV2-8            265µs ± 5%     283µs ± 4%   +6.72%  (p=0.000 n=10+10)
InitPartialResultV2/IsAuthorized(default_deny)_with_the_result_of_BenchmarkInitPartialResultV2-8     265µs ± 4%     294µs ± 8%  +11.06%  (p=0.000 n=10+10)

name                                                                                              old alloc/op   new alloc/op   delta
InitPartialResultV2/default_policies-8                                                               225MB ± 0%      24MB ± 0%  -89.52%  (p=0.000 n=10+10)
InitPartialResultV2/IsAuthorized(true)_with_the_result_of_BenchmarkInitPartialResultV2-8            89.1kB ± 0%    98.0kB ± 0%  +10.04%  (p=0.000 n=9+9)
InitPartialResultV2/IsAuthorized(false)_with_the_result_of_BenchmarkInitPartialResultV2-8           87.1kB ± 0%    95.6kB ± 0%   +9.68%  (p=0.000 n=9+10)
InitPartialResultV2/IsAuthorized(default_deny)_with_the_result_of_BenchmarkInitPartialResultV2-8    87.1kB ± 0%    95.6kB ± 0%   +9.68%  (p=0.000 n=8+10)

name                                                                                              old allocs/op  new allocs/op  delta
InitPartialResultV2/default_policies-8                                                               7.25M ± 0%     0.91M ± 0%  -87.50%  (p=0.000 n=10+9)
InitPartialResultV2/IsAuthorized(true)_with_the_result_of_BenchmarkInitPartialResultV2-8             2.24k ± 0%     2.49k ± 0%  +11.12%  (p=0.000 n=10+10)
InitPartialResultV2/IsAuthorized(false)_with_the_result_of_BenchmarkInitPartialResultV2-8            2.17k ± 0%     2.41k ± 0%  +10.72%  (p=0.000 n=10+10)
InitPartialResultV2/IsAuthorized(default_deny)_with_the_result_of_BenchmarkInitPartialResultV2-8     2.17k ± 0%     2.41k ± 0%  +10.72%  (p=0.000 n=10+10)
```

ℹ️ **Note** how we have a slight increase for `V2IsAuthorized` calls -- but a giant decrease for `InitPartialResultV2`. Let's see how it works out, and if this is a good compromise.

#### What's up with the inlining?

Using an `opa` binary built from master (the latest release, 0.12.0, doesn't have this code yet), we can examine the partial eval results for IAM v2 given our default policies (working directory `components/authz-service/engine/opa/policy`):

👉 partial results before  👈 

```
$ opa eval -d ../example_v2/store-pretty.json --partial -d authz_v2.rego -d common.rego --format pretty data.authz_v2.authorized
+-----------+-------------------------------------------------------------------------------------------+
| Query 1   | data.partial.authz_v2.authorized                                                          |
+-----------+-------------------------------------------------------------------------------------------+
| Support 1 | package partial                                                                           |
|           |                                                                                           |
|           | __not1_1__ {                                                                              |
|           |   "iam:policies:administrator-access" = input.resource                                    |
|           |   "iam:policies:update" = input.action                                                    |
|           |   _ = input.subjects[_]                                                                   |
|           | }                                                                                         |
|           |                                                                                           |
|           | __not1_1__ {                                                                              |
|           |   "iam:policies:administrator-access" = input.resource                                    |
|           |   "iam:policies:delete" = input.action                                                    |
|           |   _ = input.subjects[_]                                                                   |
|           | }                                                                                         |
|           |                                                                                           |
|           | __not1_1__ {                                                                              |
|           |   "iam:policies:editor-access" = input.resource                                           |
|           |   "iam:policies:update" = input.action                                                    |
|           |   _ = input.subjects[_]                                                                   |
|           | }                                                                                         |
|           |                                                                                           |
| ......... |    [ a few dozen more ]                                                                   |
|           |                                                                                           |
+-----------+-------------------------------------------------------------------------------------------+
| Support 2 | package partial.authz_v2                                                                  |
|           |                                                                                           |
|           | authorized {                                                                              |
|           |   startswith(input.subjects[_], "user:")                                                  |
|           |   split(input.action, ":") = ["infra", "nodeManagers", _]                                 |
|           |                                                                                           |
|           |   not data.partial.__not1_1__                                                             |
|           |   _ = input.resource                                                                      |
|           | }                                                                                         |
|           |                                                                                           |
|           | authorized {                                                                              |
|           |   "iam:policyVersion" = input.resource                                                    |
|           |   "iam:policies:get" = input.action                                                       |
|           |                                                                                           |
|           |   not data.partial.__not1_1__                                                             |
|           |   _ = input.subjects[_]                                                                   |
|           | }                                                                                         |
|           |                                                                                           |
|           | authorized {                                                                              |
|           |   "iam:introspect" = input.resource                                                       |
|           |   split(input.action, ":") = ["iam", "introspect", _]                                     |
|           |                                                                                           |
|           |   not data.partial.__not1_1__                                                             |
|           |   _ = input.subjects[_]                                                                   |
|           | }                                                                                         |
|           |                                                                                           |
| ......... |    [ a few dozen more ]                                                                   |
|           |                                                                                           |
|           | default authorized = false                                                                |
+-----------+-------------------------------------------------------------------------------------------+
```
</details>

As you can see, all `deny` rules are collected into a new support rule, `__not1_1__`, and `not data.partial.__not1_1__` is can be found in every allow rule -- it's _inlined_. (This probably happens for every `allow` rule, so, it's quadratic. Don't pin me down on the details here. :wink:)

With the new upstream feature, we can disable that:

👉 partial results without inlining data.authz_v2.deny  👈 
```
$ opa eval -d ../example_v2/store-pretty.json --partial -d authz_v2.rego -d common.rego --format pretty --disable-inlining data.authz_v2.deny data.authz_v2.authorized
+-----------+------------------------------------------------------------------------------------------+
| Query 1   | data.partial.authz_v2.authorized                                                         |
+-----------+------------------------------------------------------------------------------------------+
| Support 1 | package partial.authz_v2                                                                 |
|           |                                                                                          |
|           | deny {                                                                                   |
|           |   startswith(input.subjects[_], "user:local:")                                           |
|           |   split(input.subjects[_], ":", ["user", _, username39])                                 |
|           |   replace("iam:users:${a2:username}", "${a2:username}", username39, input.resource)      |
|           |   "iam:users:delete" = input.action                                                      |
|           | }                                                                                        |
|           |                                                                                          |
|           | deny {                                                                                   |
|           |   "iam:roles:owner" = input.resource                                                     |
|           |   "iam:roles:update" = input.action                                                      |
|           |   _ = input.subjects[_]                                                                  |
|           | }                                                                                        |
|           |                                                                                          |
|           | deny {                                                                                   |
|           |   "iam:roles:owner" = input.resource                                                     |
|           |   "iam:roles:delete" = input.action                                                      |
|           |   _ = input.subjects[_]                                                                  |
|           | }                                                                                        |
|           |                                                                                          |
| ......... |    [ a few dozen more ]                                                                  |
|           |                                                                                          |
|           | authorized {                                                                             |
|           |   startswith(input.subjects[_], "user:local:")                                           |
|           |   split(input.subjects[_], ":", ["user", _, username18])                                 |
|           |   replace("iam:users:${a2:username}", "${a2:username}", username18, input.resource)      |
|           |   "iam:users:get" = input.action                                                         |
|           |   not data.partial.authz_v2.deny                                                         |
|           | }                                                                                        |
|           |                                                                                          |
|           | authorized {                                                                             |
|           |   startswith(input.subjects[_], "user:local:")                                           |
|           |   split(input.subjects[_], ":", ["user", _, username18])                                 |
|           |   replace("iam:users:${a2:username}", "${a2:username}", username18, input.resource)      |
|           |   "iam:users:list" = input.action                                                        |
|           |   not data.partial.authz_v2.deny                                                         |
|           | }                                                                                        |
|           |                                                                                          |
|           | authorized {                                                                             |
|           |   startswith(input.subjects[_], "user:local:")                                           |
|           |   split(input.subjects[_], ":", ["user", _, username18])                                 |
|           |   replace("iam:users:${a2:username}", "${a2:username}", username18, input.resource)      |
|           |   "iam:users:update" = input.action                                                      |
|           |   not data.partial.authz_v2.deny                                                         |
|           | }                                                                                        |
|           |                                                                                          |
| ......... |    [ a few dozen more ]                                                                  |
|           |                                                                                          |
|           | default authorized = false                                                               |
+-----------+------------------------------------------------------------------------------------------+
```
</details>

Here, you can see that there's no `allow` rules, they **have been inlined** into `authorized`; but `deny` is kept as-is. For our policy logic, that's pretty much results in the same partial eval state, but it's a lot less work to compute it.

On the command line, the performance difference is noticeable, too:
```
$ opa eval -d ../example_v2/store-pretty.json --partial -d authz_v2.rego -d common.rego --format pretty --metrics data.authz_v2.authorized  | tail -10
+------------------------------+-----------+
|            METRIC            |   VALUE   |
+------------------------------+-----------+
| timer_rego_input_parse_ns    | 504       |
| timer_rego_module_compile_ns | 8561705   |
| timer_rego_module_parse_ns   | 30373423  |
| timer_rego_partial_eval_ns   | 354771125 |
| timer_rego_query_compile_ns  | 55901     |
| timer_rego_query_parse_ns    | 193525    |
+------------------------------+-----------+
$ opa eval -d ../example_v2/store-pretty.json --partial -d authz_v2.rego -d common.rego --format pretty --disable-inlining data.authz_v2.deny --metrics data.authz_v2.authorized | tail -10
+------------------------------+----------+
|            METRIC            |  VALUE   |
+------------------------------+----------+
| timer_rego_input_parse_ns    | 413      |
| timer_rego_module_compile_ns | 8414356  |
| timer_rego_module_parse_ns   | 30116282 |
| timer_rego_partial_eval_ns   | 20039244 |
| timer_rego_query_compile_ns  | 53213    |
| timer_rego_query_parse_ns    | 153601   |
+------------------------------+----------+
```
☝️ Note the `timer_rego_partial_eval_ns` metric.


### :+1: Definition of Done

- It's less annoying to create v2 policies.

### :athletic_shoe: Demo Script / Repro Steps

- deployinate, `rebuild components/authz-service`

### :chains: Related Resources

- [A2-798](https://chefio.atlassian.net/browse/A2-798)
- https://github.com/open-policy-agent/opa/issues/1447

### :white_check_mark: Checklist

- [X] Necessary tests added/updated?
- [ ] Necessary docs added/updated?
- [X] Code actually executed?
- [X] Vetting performed (unit tests, lint, etc.)?